chore: drop Robolectric in favor of plain fakes and instrumented tests
Robolectric only ever stood in for a working Context/Uri/ContentResolver
on the JVM. Tests that only needed an inert Context (filesDir, getString)
now use hand-written fakes (FakeContext, FakeFolderPathManager); tests
that genuinely need a real android.net.Uri, FileProvider, ContentResolver,
Intent, or SDK-version branching move to androidTest, since Uri can't be
faked (package-private constructor) and Robolectric was the only thing
making that work on the JVM.
