--- app/src/test/java/com/grappim/hateitorrateit/ui/screens/main/NonFirebaseApplication.kt
@@ -1,5 +1,0 @@
-package com.grappim.hateitorrateit.ui.screens.main
-
-import android.app.Application
-
-class NonFirebaseApplication : Application()
