+++ feature/home/ui/build.gradle.kts
@@ -29,6 +29,4 @@ dependencies {
     implementation(libs.androidx.hilt.navigation.compose)
 
     implementation(libs.coil)
-
-    testImplementation(libs.robolectric)
 }
