+++ feature/settings/ui/src/test/java/com/grappim/hateitorrateit/feature/settings/ui/optionsgenerator/LocaleOptionsGeneratorImplTest.kt
@@ -4,13 +4,8 @@ import com.grappim.hateitorrateit.strings.RString
 import com.grappim.hateitorrateit.utils.ui.NativeText
 import org.junit.Before
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.Config
 import kotlin.test.assertEquals
 
-@RunWith(RobolectricTestRunner::class)
-@Config(manifest = Config.NONE)
 class LocaleOptionsGeneratorImplTest {
 
     private lateinit var localeOptionsGenerator: LocaleOptionsGeneratorImpl
