+++ testing/core/build.gradle.kts
@@ -14,7 +14,6 @@ dependencies {
 
     api(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.navigation.compose)
-    implementation(libs.robolectric)
 
     api(libs.junit4)
     api(libs.kotlinx.coroutines.test)
