+++ testing/core/src/main/java/com/grappim/hateitorrateit/testing/core/FakeContext.kt
@@ -1,0 +1,11 @@
+package com.grappim.hateitorrateit.testing.core
+
+import android.content.ContextWrapper
+import java.io.File
+import java.nio.file.Files
+
+class FakeContext(
+    private val filesDirOverride: File = Files.createTempDirectory("fake-context").toFile()
+) : ContextWrapper(null) {
+    override fun getFilesDir(): File = filesDirOverride
+}
